Changed sound lengths. Used formula sound_length + sound_length * speed / (max_speed - min_speed -1) for duration of sounds. Not sure why this works, as there isn't supposed to be a -1 in the denominator. Changed the colour of the arrow in the combo boxes to be white. Added code to make wheel stop spinning when the stop button is clicked. Previous code was not working in web start. Fixed the XML writer to release lock on file once finished saving it. Changed invocation code to create UI on the swing thread. Deleted the secondary sounds folder. Both applet and application version of code use sounds in the src directory.
git-svn-id: https://subversion.hss.rpi.edu/svn/csdt_src/RhythmWheels@1133 c6ea07bc-8d12-184c-9ef1-647ee94933e0

+++ src/rhythmwheels/Painter.java
@@ -137,8 +137,8 @@ class Painter implements ActionListener
                 wheelIterations[i] = 0;
             }
             numsounds = wheels[i].getSounds().size();
-            soundLength[i] = numsounds *
-                    (Sound.SOUND_LENGTH + Sound.SOUND_LENGTH * cp.getSpeed()/(cp.MAX_SPEED - cp.MIN_SPEED));
+            soundLength[i] = (long)(numsounds *
+                    (Sound.SOUND_LENGTH + Sound.SOUND_LENGTH * cp.getSpeed()/(double)(cp.MAX_SPEED - cp.MIN_SPEED)));
             wheels[i].setSoundsPlayedCounter(0);
             wheels[i].setPreviousRotationAngle(0);
